requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
requests==2.32.3
PyJWT==2.9.0
celery==5.4.0
redis==5.0.7
